+++ CHANGELOG.md
@@ -7,9 +7,11 @@
 
 * NEW: Added full whois.audns.net.au parser.
 
-* NEW: Added full whois.cctld.by server (GH-154).
+* NEW: Added full whois.cctld.by parser (GH-154).
 
-* NEW: Added full whois.domainregistry.ie server.
+* NEW: Added full whois.domainregistry.ie parser.
+
+* NEW: Added full whois.cira.ca parser.
 
 * CHANGED: Moved scanners from Whois::Record::Parser::Scanners to Whois::Record::Scanners.